broke the sql tables into multiple files and updated imports
