+++ newsbot/sql/exceptions.py
@@ -1,0 +1,7 @@
+
+
+class FailedToAddToDatabase(Exception):
+    pass
+
+class FailedToAddRecord(Exception):
+    pass
